+++ Cesium3DTilesSelection/src/TilesetContentManager.cpp
@@ -689,7 +689,7 @@ TilesetContentManager::TilesetContentManager(
                 result.errors.emplaceError(fmt::format(
                     "Error when parsing tileset JSON, error code {} at byte "
                     "offset {}",
-                    tilesetJson.GetParseError(),
+                    (int)tilesetJson.GetParseError(),
                     tilesetJson.GetErrorOffset()));
                 return asyncSystem.createResolvedFuture(std::move(result));
               }
